Crate geometry3d

Modules

bbox3d

cylinder3d

disk3d

distant_source3d

intersection

loop3d

plane3d

point3d

polygon3d

ray3d

segment3d

sphere3d

transform

triangle3d

triangulation3d

utils

vector3d

Macros

gamma

Structs

BBox3D

An Axis-aligned Bounding Box, represented by two corners (i.e., by two Point3D objects)

Cylinder3D

A Cylinder of radius radius aligned witht the Z axis, starting at a Z value of zmin and ending at zmax.

Disk3D

A disk of radius radius whose normal points in the normal direction, located at a height height. Can have a

DistantSource3D

Represents a solid angle pointing from any Point3D in the scene towards a certain direction (i.e., a Vector3D). So, this is not really a surface because it is at an infinite distance from any point.

Loop3D

Plane3D

Describes a plane based on a Point3D (P) contained in the plane and a Vector3D (N) normal to the plane.

Point3D

Polygon3D

Ray3D

Segment3D

An imaginary line starting at one Point3D and ending on another Point3D.

Sphere3D

A Full or Partial sphere in three dimensions, centered at the origin

Transform

Represents a transformation matrix

Triangle3D

Triangulation3D

A set of Triangle3D that, together, cover completely a Polygon3D.

Vector3D

Enums

BBoxAxis

Represents one dimension of the BBox3D

PointInTriangle

Constants

PI

Type Definitions

Float

RefCount