Struct geometry3d::disk3d::Disk3D

pub struct Disk3D {
    centre: Point3D,
    normal: Vector3D,
    radius: f64,
    inner_radius: f64,
    phi_zero: Vector3D,
    phi_max: f64,
    transform: Option<Rc<Transform>>,
}

A disk of radius radius whose normal points in the normal direction, located at a height height. Can have a

Fields
centre: Point3D

The location of the center of the Disk3D

normal: Vector3D

The normal of the Disk3D

radius: f64

The external radius of the Disk3D

inner_radius: f64

The radius of the inner hole of the Disk3D

phi_zero: Vector3D

The axis that marks phi as Zero

phi_max: f64

The angle described by the Disk3D. A value of 2.*PI implies a full Disk3D; a value of less than that, a partial Disk3D.

transform: Option<Rc<Transform>>

A pointer to the Transform associated with this Disk3D

Implementations

impl Disk3D

pub fn new(centre: Point3D, normal: Vector3D, radius: f64) -> Self

Creates a new Disk3D based on its centre, normal and radius.

pub fn new_detailed(
centre: Point3D,
normal: Vector3D,
radius: f64,
inner_radius: f64,
phi_zero: Vector3D,
phi_max: f64,
transform: Option<Rc<Transform>>
) -> Self

Creates a new—potentially partial—Disk3D

pub fn basic_intersection(
&self,
ray: &Ray3D,
_o_error: Point3D,
_d_error: Point3D
) -> Option<(Point3D, f64)>

Intersects a Ray3D in local coordinates with the Disk3D. Returns the a Point3D of intersection and phi, the parameter indicating the polar position of the point of intersection.

pub fn intersection_info(
&self,
ray: &Ray3D,
phit: Point3D,
phi: f64
) -> Option<IntersectionInfo>

Intersects a ray—in world coordinates—and a Disk3D, returning the IntersectionInfo

pub fn bounds(&self) -> BBox3D

Gets a BBox3D bounding the object, in local coordinates

pub fn id(&self) -> &'static str

The name of the figure. Useful for debugging.

pub fn area(&self) -> f64

Gets the area of the object

pub fn transform(&self) -> &Option<Rc<Transform>>

Borrows the Transform

pub fn simple_intersect_local_ray(
&self,
ray: &Ray3D,
o_error: Point3D,
d_error: Point3D
) -> Option<Point3D>

Like intersect_local_ray but simplified because there is not need for calcuating the paramtrisized elements

pub fn intersect_local_ray(
&self,
ray: &Ray3D,
o_error: Point3D,
d_error: Point3D
) -> Option<IntersectionInfo>

Intersects a ray—in local coordinates—and a Disk3D, returning the IntersectionInfo

pub fn intersect(&self, ray: &Ray3D) -> Option<IntersectionInfo>

Intersects an object with a [Ray3D] (IN WORLD COORDINATES) traveling forward, returning a detailed IntersectionInfo about the intersaction .

pub fn simple_intersect(&self, ray: &Ray3D) -> Option<Point3D>

Intersects an object with a [Ray3D] (IN WORLD COORDINATES) traveling forward, returning the point of intersection, if any.

pub fn world_bounds(&self) -> BBox3D

Gets a BBox3D bounding the object, in world’s coordinates.