Struct geometry3d::distant_source3d::DistantSource3D

pub struct DistantSource3D {
    pub direction: Vector3D,
    pub omega: f64,
    pub angle: f64,
    cos_half_alpha: f64,
    tan_half_alpha: f64,
    transform: Option<Rc<Transform>>,
}

Represents a solid angle pointing from any Point3D in the scene towards a certain direction (i.e., a Vector3D). So, this is not really a surface because it is at an infinite distance from any point.

Fields
direction: Vector3D

The direction pointing to the souRefCounte

omega: f64

The solid angle in Sr

angle: f64

The angle in Radians

cos_half_alpha: f64

Cos half angle

tan_half_alpha: f64

Tan of half angle

transform: Option<Rc<Transform>>

A pointer to the Transform associated with this [DistantSouRefCounte3D]

Implementations

impl DistantSource3D

pub fn new(direction: Vector3D, angle: f64) -> Self

Creates a new [DistantSouRefCounte3D] geometry.

Inputs:

direction: A Vector3D pointing to the souRefCounte

angle: The flat angle in Radians (e.g., a hemisphere would be $\pi$)

pub fn get_proxy_disk(&self, t: f64) -> Disk3D

Creates a proxy Disk3D that can be used for sampling and other operations

pub fn id(&self) -> &'static str

The name of the figure. Useful for debugging.

pub fn bounds(&self) -> BBox3D

Gets a BBox3D bounding the object, in local coordinates

pub fn area(&self) -> f64

Gets the area of the object

pub fn transform(&self) -> &Option<Rc<Transform>>

Borrows the Transform

pub fn intersect_local_ray(
&self,
ray: &Ray3D,
o_error: Point3D,
d_error: Point3D
) -> Option<IntersectionInfo>

Intersects an object with a [Ray3D] (IN LOCAL COORDINATES) traveling forward, returning the distance t and the normal Vector3D at that point. If the distance is negative (i.e., the object is behind the plane), it should return None.

The steps should be:

Calculate the point of intersection

Calculate normal and side of intersection

Calculate (u,v)

Calculate first derivative (i.e., dp_du, dp_dv)

Calculate second derivative (i.e., d2p_duu, d2p_dvv, d2p_duv)

Call IntersectionInfo::new(ray,p, u, v, dpdu, dpdv, d2p_duu, d2p_dvv, d2p_duv)

pub fn simple_intersect_local_ray(
&self,
ray: &Ray3D,
_o_error: Point3D,
_d_error: Point3D
) -> Option<Point3D>

Like intersect_local_ray but simplified because there is not need for calcuating the paramtrisized elements

pub fn intersect(&self, ray: &Ray3D) -> Option<IntersectionInfo>

Intersects an object with a [Ray3D] (IN WORLD COORDINATES) traveling forward, returning the distance t and the normal Vector3D at that point. If the distance is negative (i.e., the object is behind the plane), it should return None. Returns detailed IntersectionInfo about the intersaction .

pub fn simple_intersect(&self, ray: &Ray3D) -> Option<Point3D>

Intersects an object with a [Ray3D] (IN WORLD COORDINATES) traveling forward, returning the distance t and the normal Vector3D at that point. If the distance is negative (i.e., the object is behind the plane), it should return None. Returns only the point of intersection.

pub fn world_bounds(&self) -> BBox3D

Gets a BBox3D bounding the object, in world’s coordinates.