Trait geometry3d::intersect_trait::Intersect

pub trait Intersect {
    fn id(&self) -> &'static str;

    fn area(&self) -> f64;

    fn transform(&self) -> &Option<Rc<Transform>>;

    fn intersect_local_ray(
        &self, 
        ray: &Ray3D, 
        o_error: Point3D, 
        d_error: Point3D
    ) -> Option<IntersectionInfo>;

    fn simple_intersect_local_ray(
        &self, 
        ray: &Ray3D, 
        o_error: Point3D, 
        d_error: Point3D
    ) -> Option<Point3D>;

    fn intersect(&self, ray: &Ray3D) -> Option<IntersectionInfo> { ... }

    fn simple_intersect(&self, ray: &Ray3D) -> Option<Point3D> { ... }
}

Required methods

fn id(&self) -> &'static str

The name of the figure. Useful for debugging.

fn area(&self) -> f64

Gets the area of the object

fn transform(&self) -> &Option<Rc<Transform>>

Borrows the Transform

fn intersect_local_ray(
&self,
ray: &Ray3D,
o_error: Point3D,
d_error: Point3D
) -> Option<IntersectionInfo>

Intersects an object with a [Ray3D] (IN LOCAL COORDINATES) traveling forward, returning the distance t and the normal Vector3D at that point. If the distance is negative (i.e., the object is behind the plane), it should return None.

The steps should be:

Calculate the point of intersection

Calculate normal and side of intersection

Calculate (u,v)

Calculate first derivative (i.e., dp_du, dp_dv)

Calculate second derivative (i.e., d2p_duu, d2p_dvv, d2p_duv)

Call IntersectionInfo::new(ray,p, u, v, dpdu, dpdv, d2p_duu, d2p_dvv, d2p_duv)

fn simple_intersect_local_ray(
&self,
ray: &Ray3D,
o_error: Point3D,
d_error: Point3D
) -> Option<Point3D>

Like intersect_local_ray but simplified because there is not need for calcuating the paramtrisized elements

Provided methods

fn intersect(&self, ray: &Ray3D) -> Option<IntersectionInfo>

Intersects an object with a [Ray3D] (IN WORLD COORDINATES) traveling forward, returning the distance t and the normal Vector3D at that point. If the distance is negative (i.e., the object is behind the plane), it should return None. Returns detailed IntersectionInfo about the intersaction .

fn simple_intersect(&self, ray: &Ray3D) -> Option<Point3D>

Intersects an object with a [Ray3D] (IN WORLD COORDINATES) traveling forward, returning the distance t and the normal Vector3D at that point. If the distance is negative (i.e., the object is behind the plane), it should return None. Returns only the point of intersection.