Struct geometry3d::intersection::IntersectionInfo

pub struct IntersectionInfo {
    pub p: Point3D,
    pub normal: Vector3D,
    pub side: SurfaceSide,
    pub dpdu: Vector3D,
    pub dpdv: Vector3D,
    pub dndu: Vector3D,
    pub dndv: Vector3D,
    pub u: f64,
    pub v: f64,
}

Contains more detailed information about the what is happening at the surface in the intersection point

Fields
p: Point3D

The point of intersection

normal: Vector3D

The normal Vector3D at the interaction. Must have the value of (dpdu x dpdv).normalize() … use ShadingInfo::new(..) and this is done automatically

side: SurfaceSide

The side in which the ray intersected

dpdu: Vector3D

Partial derivative of the position p with respect to u

dpdv: Vector3D

Partial derivative of the position p with respect to v

dndu: Vector3D

Partial derivative of the normal n with respect to u

dndv: Vector3D

Partial derivative of the normal n with respect to v

u: f64

The position u of the intersection point

v: f64

The position v of the intersection point

Implementations

impl IntersectionInfo

pub fn new(
ray: &Ray3D,
p: Point3D,
u: f64,
v: f64,
dpdu: Vector3D,
dpdv: Vector3D,
d2p_duu: Vector3D,
d2p_dvv: Vector3D,
d2p_duv: Vector3D
) -> Self

Creates a new IntersectionInfo object

pub fn transform(&self, transform: &Transform) -> Self

Applies a Transform to the IntersectionInfo, creating a new one.

pub fn inv_transform(&self, transform: &Transform) -> Self

Applies the inverse of a Transform to the IntersectionInfo, creating a new one.