Struct geometry3d::intersection_no_texture::IntersectionInfo

pub struct IntersectionInfo {
    pub p: Point3D,
    pub normal: Vector3D,
    pub side: SurfaceSide,
    pub dpdu: Vector3D,
    pub dpdv: Vector3D,
}

Contains more detailed information about the what is happening at the surface in the intersection point

Fields
p: Point3D

The point of intersection

normal: Vector3D

The normal Vector3D at the interaction. Must have the value of (dpdu x dpdv).normalize() … use ShadingInfo::new(..) and this is done automatically

side: SurfaceSide

The side in which the ray intersected

dpdu: Vector3D

Partial derivative of the position p with respect to u

dpdv: Vector3D

Partial derivative of the position p with respect to v

Implementations

impl IntersectionInfo

pub fn new(
ray: &Ray3D,
p: Point3D,
_u: f64,
_v: f64,
dpdu: Vector3D,
dpdv: Vector3D,
_d2p_duu: Vector3D,
_d2p_dvv: Vector3D,
_d2p_duv: Vector3D
) -> Self

pub fn transform(&self, transform: &Transform) -> Self

pub fn inv_transform(&self, transform: &Transform) -> Self