Module geometry3d::round_error

Rounding Error accumulation module, based on PBRT

Structs

ApproxFloat

Contains a Float value as well as an error interval.

Constants

MINUS_ZERO 🔒

ZERO 🔒

Functions

max_min

Calculates the minimum and the maximum of a &[Float; 4]

next_float_down 🔒

next_float_up 🔒