Struct geometry3d::BBox3D

pub struct BBox3D {
    pub min: Point3D,
    pub max: Point3D,
}

An Axis-aligned Bounding Box, represented by two corners (i.e., by two Point3D objects)

Fields
min: Point3D

The lower corner

max: Point3D

Higher corner

Implementations

impl BBox3D

pub fn new(a: Point3D, b: Point3D) -> Self

Creates a new BBox3D from two points. The points will be all scrambled up in order to find the minimum x, y and z.

pub fn from_point(p: Point3D) -> Self

Creates a BBox3D that contains a single point, thus its min and its max values are equal

pub fn from_union_point(bbox: &Self, pt: Point3D) -> Self

Creates a new BBox3D that contains an existing BBox3D and a Point3D

pub fn from_union(bbox1: &Self, bbox2: &Self) -> Self

Creates a new BBox3D that contains two existing BBox3D

pub fn from_intersection(bbox1: &Self, bbox2: &Self) -> Self

Creates a new BBox3D that contains the intersection of two existing BBox3D

pub fn overlaps(&self, other: &Self) -> bool

Checks if a BBox3D overlaps with another one

pub fn point_inside(&self, pt: Point3D) -> bool

Checks if a Point3D is inside this BBox3D

pub fn point_inside_exclusive(&self, pt: Point3D) -> bool

Checks if a Point3D is inside this BBox3D but any Point3D located in the upper boundary is considered to be outside