Struct geometry3d::Cylinder3D

pub struct Cylinder3D {
    radius: f64,
    zmin: f64,
    zmax: f64,
    phi_max: f64,
    transform: Option<Rc<Transform>>,
}

A Cylinder of radius radius aligned witht the Z axis, starting at a Z value of zmin and ending at zmax.

It can be an ‘open’ cylinder, by setting a phi_max less than 2.*PI

Fields
radius: f64

The radius of the Cylinder3D

zmin: f64

The Z coordinate of the lower end of the Cylinder3D

zmax: f64

The Z coordinate of the upper end of the Cylinder3D

phi_max: f64

The range of longitudes allowed (2*PI radians in a full Cylinder3D, less in a partial one)

transform: Option<Rc<Transform>>

A pointer to the Transform associated with this Cylinder3D

Implementations

impl Cylinder3D

pub fn new(p0: Point3D, p1: Point3D, radius: f64) -> Self

Creates a new full Cylinder3D; i.e., it is a revolution of 360 degrees

pub fn new_partial(p0: Point3D, p1: Point3D, radius: f64, phi_max: f64) -> Self

Creates a new partial Cylinder3D; i.e., it does not have to be closed. For instance, it can only be a 180 degree revolution

pub fn new_transformed(
radius: f64,
zmin: f64,
zmax: f64,
phi_max: f64,
transform: Option<Rc<Transform>>
) -> Self

The angle phi_max is in degrees

pub fn basic_intersection(
&self,
ray: &Ray3D,
o_error: Point3D,
d_error: Point3D
) -> Option<(Point3D, f64)>

Performs a basic intersection of a ray and a Cylinder3D, returning the point of intersection and the angle phi of that intersection

pub fn intersection_info(
&self,
ray: &Ray3D,
phit: Point3D,
phi: f64
) -> Option<IntersectionInfo>

Grabs the result of the basic_intersection and creates an IntersectionInfo

pub fn bounds(&self) -> BBox3D

Gets a BBox3D bounding the object, in local coordinates

pub fn id(&self) -> &'static str

The name of the figure. Useful for debugging.

pub fn area(&self) -> f64

Gets the area of the object

pub fn transform(&self) -> &Option<Rc<Transform>>

Borrows the Transform

pub fn simple_intersect_local_ray(
&self,
ray: &Ray3D,
o_error: Point3D,
d_error: Point3D
) -> Option<Point3D>

Like intersect_local_ray but simplified because there is not need for calcuating the paramtrisized elements

pub fn intersect_local_ray(
&self,
ray: &Ray3D,
o_error: Point3D,
d_error: Point3D
) -> Option<IntersectionInfo>

Intersects an object with a [Ray3D] (IN LOCAL COORDINATES) traveling forward, returning the distance t and the normal Vector3D at that point. If the distance is negative (i.e., the object is behind the plane), it should return None.

The steps should be:

Calculate the point of intersection

Calculate normal and side of intersection

Calculate (u,v)

Calculate first derivative (i.e., dp_du, dp_dv)

Calculate second derivative (i.e., d2p_duu, d2p_dvv, d2p_duv)

Call IntersectionInfo::new(ray,p, u, v, dpdu, dpdv, d2p_duu, d2p_dvv, d2p_duv)

pub fn intersect(&self, ray: &Ray3D) -> Option<IntersectionInfo>

Intersects an object with a [Ray3D] (IN WORLD COORDINATES) traveling forward, returning the distance t and the normal Vector3D at that point. If the distance is negative (i.e., the object is behind the plane), it should return None. Returns detailed IntersectionInfo about the intersaction .

pub fn simple_intersect(&self, ray: &Ray3D) -> Option<Point3D>

Intersects an object with a [Ray3D] (IN WORLD COORDINATES) traveling forward, returning the point of intersection, if any.

pub fn world_bounds(&self) -> BBox3D

Gets a BBox3D bounding the object, in world’s coordinates.