Struct geometry3d::Loop3D

pub struct Loop3D {
    vertices: Vec<Point3D>,
    normal: Vector3D,
    closed: bool,
    area: f64,
}

A set of Point3D in sequence, forming a closed loop. It has some particularities.

use geometry3d::{Loop3D, Point3D};
let mut the_loop = Loop3D::new();
assert!(the_loop.is_empty());
let l = 0.5;

the_loop.push(Point3D::new(-l, -l, 0.)).unwrap();
the_loop.push(Point3D::new(-l, l, 0.)).unwrap();
the_loop.push(Point3D::new(l, l, 0.)).unwrap();
the_loop.push(Point3D::new(l, -l, 0.)).unwrap();

assert!(the_loop.area().is_err());
the_loop.close().unwrap();

let a = the_loop.area().unwrap();
assert!((4. * l * l - a).abs() < 0.0001);

Note:

It has some peculiarities. For instance, it attempts to reduce the number of points on a Loop3D. This is done by identifying when a colinear Point3D is to be added and, instead of extending the Loop3D, replacing the last element.

use geometry3d::{Loop3D, Point3D};
let mut the_loop = Loop3D::new();
 
the_loop.push(Point3D::new(0., 0., 0.)).unwrap();
the_loop.push(Point3D::new(1., 1., 0.)).unwrap();
assert_eq!(2, the_loop.n_vertices());
 
// Adding a collinear point will not extend.
let collinear = Point3D::new(2., 2., 0.);
the_loop.push(collinear).unwrap();
assert_eq!(2, the_loop.n_vertices());
assert_eq!(the_loop[1], collinear);

Fields
vertices: Vec<Point3D>

The points of the Loop3D

normal: Vector3D

The normal, following a right-hand-side convention

closed: bool

A flag indicating whether the Loop3D is considered finished or not.

area: f64

The area of the Loop3D, only calculated when closing it

Implementations

impl Loop3D

pub fn new() -> Loop3D

Creates a new and empty Loop3D. It has a Zero Normal and an area of -1. These attributes are filled automatically when pushing vertices into the Loop.

pub fn is_empty(&self) -> bool

Checks if the Loop3D has Zero vertices

pub fn vertices(&self) -> &[Point3D]

Borrows the vertices

pub fn remove(&mut self, i: usize)

Removes a vertex

fn valid_to_add(&self, point: Point3D) -> Result<(), String>

Checks whether a Point3D can be added to a Loop3D while keeping it valid

It checks that:

That the Loop3D has not been closed yet

that the new Point3D is coplanar witht he rest of the Point3D

Adding the Point3D will not make the Loop3D intersect with itself

pub fn push(&mut self, point: Point3D) -> Result<(), String>

Pushes a new Point3D into the Loop3D.

If the Point3D being pushed is collinear with the previous two, then instead of pushing a new Point3D it will update the last one (i.e., because the shape and area) of the Loop3D will still be the same.

Returns an error if the point being added would make the Loop3D intersect itself, or if the new Point3D is not coplanar with the Loop3D, or if the Loop3D is closed.

pub fn n_vertices(&self) -> usize

Counts the vertices in the Loop3D

pub fn is_diagonal(&self, s: Segment3D) -> bool

Checks if a Segment3D intersects any of the other Segment3D in the Loop3D and if its midpoint is inside of it.

Note that the idea is to call it by using segments that go from one vertex to another. I mean, this function takes any segment, meaning that a small Segment3D “floating” inside of a big Loop3D will be considered a diagonal… be careful with this

pub fn open(&mut self)

Opens a Loop3D

pub fn close(&mut self) -> Result<(), String>

Closes a Loop3D, calculating its area and checking the connection between the first and last vertex. If the first and the last

fn set_normal(&mut self) -> Result<(), String>

Sets the normal Vector3D for a Loop3D

pub fn normal(&self) -> Vector3D

Retrieves the normal of the vector.

Note

If the Loop3D has less than 3 vertices, then the Normal will be Vector3D(0., 0., 0.), which is the default.

pub fn is_coplanar(&self, p: Point3D) -> Result<bool, String>

Checks whether a Point3D is coplanar with the rest of the points.

pub fn test_point(&self, point: Point3D) -> Result<bool, String>

Tests whether a Point3D dwells inside of the Loop3D.

fn set_area(&mut self) -> Result<f64, String>

Calculates and caches the area of the Loop3D

pub fn area(&self) -> Result<f64, String>

Returns the area of the Loop3D

pub fn is_closed(&self) -> bool

Indicates whether the Loop3D has been closed already

pub fn contains_segment(&self, s: &Segment3D) -> bool

Checks whether the Loop3D contains a Segment3D s