Struct geometry3d::Point3D

pub struct Point3D {
    pub x: f64,
    pub y: f64,
    pub z: f64,
}

A very simple implementation of a 3D-Point

Fields
x: f64

The X component

y: f64

The Y component

z: f64

The Z component

Implementations

impl Point3D

pub fn new(x: f64, y: f64, z: f64) -> Point3D

Creates a new Point3D

let pt = Point3D::new(0., 1., 2.);

pub fn is_zero(&self) -> bool

Checks whether the Point3D is located at the origin

let pt = Point3D::new(0., 0., 0.);
assert!(pt.is_zero());

pub fn squared_distance(&self, point: Point3D) -> f64

Calculates the square of the distance between two Point3D This is faster than calculating the distance()

let a = Point3D::new(0., 0., 0.);
let b = Point3D::new(2., 0., 0.);
assert_eq!(a.squared_distance(b), 4.);

pub fn distance(&self, point: Point3D) -> f64

Calculates the distance between two Point3D

let a = Point3D::new(0., 0., 0.);
let b = Point3D::new(2., 0., 0.);
assert_eq!(a.distance(b), 2.);

pub fn compare(&self, p: Point3D) -> bool

Checks if two Point3D are sifnificantly close

let a = Point3D::new(0., 0., 0.);
let b = Point3D::new(2., 0., 0.);
assert!(!a.compare(b));
assert!(a.compare(a));

pub fn is_collinear(self, b: Point3D, c: Point3D) -> Result<bool, String>

Checks if a certain Point3D is collinear with two other Point3D. Will return an error if the two other points are the same