Struct geometry3d::Segment3D

pub struct Segment3D {
    pub start: Point3D,
    pub end: Point3D,
    pub length: f64,
}

An imaginary line starting at one Point3D and ending on another Point3D.

Fields
start: Point3D

The staring Point3D

end: Point3D

The ending Point3D

length: f64

The length

Implementations

impl Segment3D

pub fn new(a: Point3D, b: Point3D) -> Segment3D

Creates a new Segment3D

pub fn start(&self) -> Point3D

Gets the start of the Segment3D

pub fn end(&self) -> Point3D

Gets the end of the Segment3D

pub fn as_vector3d(&self) -> Vector3D

Gets a Vector3D starting at the start and ending at the end

pub fn as_reversed_vector3d(&self) -> Vector3D

Gets a Vector3D starting at the end and ending at the start

pub fn length(&self) -> f64

Gets the length of the Segment3D

pub fn compare(&self, other: &Segment3D) -> bool

Compares the Start and End Point3D of two Segment3D. This will return true if the start and end are equal(ish), even if one of the segments is reversed.

pub fn contains_point(&self, point: Point3D) -> Result<bool, String>

Checks if a Segment3D contains another Point3D.

pub fn contains(&self, input: &Segment3D) -> Result<bool, String>

Checks if a Segment3D contains another Segment3D.

pub fn get_intersection_pt(&self, input: &Segment3D) -> Option<(f64, f64)>

Checks where is it that two Segment3D intersect, returning the fraction of the caller Segment3D and the input Segment3D in which the two segments would intersect. Returns None if the caller and the input segments are in the same direction or if they dwell in different planes (i.e., they do not intercept)

Examples

use geometry3d::Point3D;
use geometry3d::Segment3D;

let vertical = Segment3D::new(Point3D::new(0., 0., -1.), Point3D::new(0., 0., 1.));
let horizontal = Segment3D::new(Point3D::new(-1., 0., 0.), Point3D::new(1., 0., 0.));
assert_eq!(horizontal.get_intersection_pt(&vertical), Some((0.5, 0.5)));

let vertical = Segment3D::new(Point3D::new(0., 0., -1.), Point3D::new(0., 0., 1.));
let horizontal = Segment3D::new(Point3D::new(0., 0., 0.), Point3D::new(1., 0., 0.));
assert_eq!(horizontal.get_intersection_pt(&vertical), Some((0.0, 0.5)));

let vertical = Segment3D::new(Point3D::new(0., 0., -1.), Point3D::new(0., 0., 1.));
let horizontal = Segment3D::new(Point3D::new(0.5, 0., 0.), Point3D::new(1., 0., 0.));
assert_eq!(horizontal.get_intersection_pt(&vertical), Some((-1., 0.5)));

pub fn intersect(&self, input: &Segment3D, output: &mut Point3D) -> bool

Checks if two Segment3D intersect each other. This returns false if one of the Segment3D barely touches the other one.

pub fn touches(&self, input: &Segment3D, output: &mut Point3D) -> bool

Checks if two Segment3D intersect each other. This returns true if one of the Segment3D barely touches the other one.

pub fn midpoint(&self) -> Point3D

Gets the midpoint of the a Segment3D