Struct geometry3d::Sphere3D

pub struct Sphere3D {
    pub radius: f64,
    zmin: f64,
    zmax: f64,
    phi_max: f64,
    delta_theta: f64,
    theta_min: f64,
    transform: Option<Rc<Transform>>,
}

A Full or Partial sphere in three dimensions, centered at the origin

Everything is in Local coordinates. Where $\phi$ describes the longitude (i.e., it goes from 0 to 360 degrees); and $\theta$ defines the latitude (i.e., it goes from 0 to 190 degrees).

Fields
radius: f64

The radius of the Sphere3D

zmin: f64

The position of the lower clipping plane on the Z axis

zmax: f64

The posiftion of the upper clipping plane on the Z axis

phi_max: f64

The range of longitudes allowed for the sphere

delta_theta: f64

The range of Latitudes comprised within the allowed zmin and zmax

theta_min: f64

The minimum latitude allowed by zmin

transform: Option<Rc<Transform>>

A pointer to the Transform associated with this Sphere3D

Implementations

impl Sphere3D

pub fn new(radius: f64, centre: Point3D) -> Self

Creates a new full Sphere3D of a certain radius and position its its centre.

If the given centre is the origin (i.e., Vector3D(0., 0., 0.)), no Transform will be attached to this Sphere3D. If this is not the case, a translate Transform will be attached to it.

pub fn centre(&self) -> Point3D

Gets the Point3D of the Sphere3D that represents its centre, after transformation.

pub fn new_transformed(radius: f64, transform: Option<Rc<Transform>>) -> Self

Creates a new full Sphere3D of a certain radius and transformed according to a Transform.

pub fn new_partial(
radius: f64,
centre: Point3D,
zmin: f64,
zmax: f64,
phi_max: f64
) -> Self

Creates a new potentially partial Sphere3D of a certain radius and position its its centre. It allows specifying the zmin and zmax as well as the phi_max (in degrees). The expected values for phi_max should go between 0 and 360 degress; while those for
zmin and zmax should go between 0 and plus or minus radius.

If the given centre is the origin (i.e., Vector3D(0., 0., 0.)), no Transform will be attached to this Sphere3D. If this is not the case, a translate Transform will be attached to it.

pub fn new_partial_transformed(
radius: f64,
zmin: f64,
zmax: f64,
phi_max: f64,
transform: Option<Rc<Transform>>
) -> Self

Creates a new potentially partial Sphere3D of a certain radius and transformed according to a Transform. It allows specifying the zmin and zmax as well as the phi_max (in degrees). The expected values for phi_max should go between 0 and 360 degress; while those for
zmin and zmax should go between 0 and plus or minus radius.

fn approx_basic_intersection(
&self,
ray: &Ray3D,
o_error: Point3D,
d_error: Point3D
) -> Option<(Point3D, f64)>

pub fn intersection_info(
&self,
ray: &Ray3D,
phit: Point3D,
phi: f64
) -> Option<IntersectionInfo>

pub fn id(&self) -> &'static str

The name of the figure. Useful for debugging.

pub fn bounds(&self) -> BBox3D

Gets a BBox3D bounding the object, in local coordinates

pub fn area(&self) -> f64

Gets the area of the object

pub fn transform(&self) -> &Option<Rc<Transform>>

Borrows the Transform

pub fn intersect_local_ray(
&self,
ray: &Ray3D,
o_error: Point3D,
d_error: Point3D
) -> Option<IntersectionInfo>

Intersects an object with a [Ray3D] (IN LOCAL COORDINATES) traveling forward, returning the distance t and the normal Vector3D at that point. If the distance is negative (i.e., the object is behind the plane), it should return None.

The steps should be:

Calculate the point of intersection

Calculate normal and side of intersection

Calculate (u,v)

Calculate first derivative (i.e., dp_du, dp_dv)

Calculate second derivative (i.e., d2p_duu, d2p_dvv, d2p_duv)

Call IntersectionInfo::new(ray,p, u, v, dpdu, dpdv, d2p_duu, d2p_dvv, d2p_duv)

pub fn simple_intersect_local_ray(
&self,
ray: &Ray3D,
o_error: Point3D,
d_error: Point3D
) -> Option<Point3D>

Like intersect_local_ray but simplified because there is not need for calcuating the paramtrisized elements

pub fn intersect(&self, ray: &Ray3D) -> Option<IntersectionInfo>

Intersects an object with a [Ray3D] (IN WORLD COORDINATES) traveling forward, returning the distance t and the normal Vector3D at that point. If the distance is negative (i.e., the object is behind the plane), it should return None. Returns detailed IntersectionInfo about the intersaction .

pub fn simple_intersect(&self, ray: &Ray3D) -> Option<Point3D>

Intersects an object with a [Ray3D] (IN WORLD COORDINATES) traveling forward, returning the distance t and the normal Vector3D at that point. If the distance is negative (i.e., the object is behind the plane), it should return None. Returns only the point of intersection.

pub fn world_bounds(&self) -> BBox3D

Gets a BBox3D bounding the object, in world’s coordinates.