Struct geometry3d::Transform

pub struct Transform {
    elements: [f64; 16],
    inv_elements: [f64; 16],
}

Represents a transformation matrix

Fields
elements: [f64; 16]inv_elements: [f64; 16]
Implementations

impl Transform

pub fn new() -> Self

Crates a new Identity Transform (i.e., does nothing to the points being passed)

pub fn translate(x: f64, y: f64, z: f64) -> Self

Creates a new Translation Transform

pub fn scale(x: f64, y: f64, z: f64) -> Self

Creates a new Scale Transform

pub fn rotate_x(degrees: f64) -> Self

Creates a new Transform that rotates counterclockwise on the X axis. The input is in degrees

pub fn rotate_y(degrees: f64) -> Self

Creates a new Transform that rotates counterclockwise on the Y axis. The input is in degrees

pub fn rotate_z(degrees: f64) -> Self

Creates a new Transform that rotates counterclockwise on the Z axis. The input is in degrees

pub fn changes_hands(&self) -> bool

Checks whether a Transform modifies the hand-ness of the coordinate system

pub fn transform_pt(&self, pt: Point3D) -> Point3D

Transforms a Point3D into another Point3D

pub fn inv_transform_pt(&self, pt: Point3D) -> Point3D

Inverse Transforms a Point3D into another Point3D

pub fn transform_pt_with_error(&self, pt: Point3D) -> (Point3D, Point3D)

Transforms a Point3D into another Point3D, propagating an error value

pub fn inv_transform_pt_with_error(&self, pt: Point3D) -> (Point3D, Point3D)

Inverse Transforms a Point3D into another Point3D, propagating an error value

pub fn transform_pt_propagate_error(
&self,
pt: Point3D,
error: Point3D
) -> (Point3D, Point3D)

Transforms a Point3D into another Point3D, calculating an error value

pub fn inv_transform_pt_propagate_error(
&self,
pt: Point3D,
error: Point3D
) -> (Point3D, Point3D)

Inverse Transforms a Point3D into another Point3D, calculating an error value

pub fn transform_vec(&self, vec: Vector3D) -> Vector3D

Transforms a Vector3D into another Vector3D

pub fn inv_transform_vec(&self, vec: Vector3D) -> Vector3D

Inverse Transforms a Vector3D into another Vector3D

pub fn transform_vec_with_error(&self, vec: Vector3D) -> (Vector3D, Point3D)

Transforms a Vector3D into another Vector3D, calculating an error value

pub fn inv_transform_vec_with_error(&self, vec: Vector3D) -> (Vector3D, Point3D)

Inverse Transforms a Vector3D into another Vector3D, calculating an error value

pub fn transform_vec_propagate_error(
&self,
vec: Vector3D,
error: Point3D
) -> (Vector3D, Point3D)

Transforms a Vector3D into another Vector3D, propagating an error value

pub fn inv_transform_vec_propagate_error(
&self,
vec: Vector3D,
error: Point3D
) -> (Vector3D, Point3D)

Inverse Transforms a Vector3D into another Vector3D, propagating an error value

pub fn transform_normal(&self, vec: Vector3D) -> Vector3D

Transforms a Vector3D into another Point3D

pub fn inv_transform_normal(&self, vec: Vector3D) -> Vector3D

Inv transforms a Vector3D into another Point3D

pub fn transform_ray(&self, ray: &Ray3D) -> (Ray3D, Point3D, Point3D)

Transforms a Ray3D. returns a new Ray3D and also the errors associated to the origin and the direction in a tuple

pub fn inv_transform_ray(&self, ray: &Ray3D) -> (Ray3D, Point3D, Point3D)

Inverse Transforms a Ray3D. returns a new Ray3D and also the errors associated to the origin and the direction in a tuple

pub fn transform_ray_propagate_error(
&self,
ray: &Ray3D,
o_error_in: Point3D,
d_error_in: Point3D
) -> (Ray3D, Point3D, Point3D)

Transforms a Ray3D and propagates the error. returns a new Ray3D and also the errors associated to the origin and the direction in a tuple

pub fn inv_transform_ray_propagate_error(
&self,
ray: &Ray3D,
o_error_in: Point3D,
d_error_in: Point3D
) -> (Ray3D, Point3D, Point3D)

Inverse Transforms a Ray3D and propagates the error. returns a new Ray3D and also the errors associated to the origin and the direction in a tuple

pub fn transform_bbox(&self, bbox: BBox3D) -> BBox3D

pub fn inv_transform_bbox(&self, bbox: BBox3D) -> BBox3D