Struct geometry3d::Triangle3D

pub struct Triangle3D {
    a: Point3D,
    b: Point3D,
    c: Point3D,
    normal: Vector3D,
    area: f64,
}

A simple 3-dimentional Triangle.

Fields
a: Point3D

First vertex

b: Point3D

Second vertex

c: Point3D

Third vertex

normal: Vector3D

The normal, following right-hand convention

area: f64

The area of the triangle

Implementations

impl Triangle3D

pub fn new(
vertex_a: Point3D,
vertex_b: Point3D,
vertex_c: Point3D
) -> Result<Self, String>

Creates a new Triangle3D

fn set_normal(&mut self)

Calculates the normal of the Triangle

pub fn normal(&self) -> Vector3D

Gets the normal of the Triangle3D. This value is cached, so this should be fast.

fn set_area(&mut self)

Sets the area of the Triangle3D. This is used when constructing the Triangle3D

pub fn area(&self) -> f64

Gets the Area of the Triangle3D… it is cached when creater

pub fn circumradius(&self) -> f64

Gets the Circumradus of the Triangle3D. This value is not cached, so it might be slow to rely on this too often.

pub fn aspect_ratio(&self) -> f64

Gets the Aspect Ratio of the Triangle3D. This is useful for triangulating polygons.

pub fn circumcenter(&self) -> Point3D

retrieves the Circumcenter

pub fn centroid(&self) -> Point3D

Gets the centroid of the Triangle3D. This value is not cached, so it might be slow to rely on this too often.

pub fn vertex(&self, i: usize) -> Result<Point3D, String>

Gets a specific vertex. Returns an error if the number of the vertex is greater than 2 (i.e., valid vertices are 0, 1, 2)

pub fn a(&self) -> Point3D

Retrieves the first vertex of an A,B,C Triangle3D

pub fn b(&self) -> Point3D

Retrieves the second vertex of an A,B,C Triangle3D

pub fn c(&self) -> Point3D

Retrieves the third vertex of an A,B,C Triangle3D

pub fn segment(&self, i: usize) -> Result<Segment3D, String>

Retrieves a segment of an A,B,C Triangle3D

pub fn ab(&self) -> Segment3D

Retrieves the first segment of an A,B,C Triangle3D

pub fn bc(&self) -> Segment3D

Retrieves the second segment of an A,B,C Triangle3D

pub fn ca(&self) -> Segment3D

Retrieves the third segment of an A,B,C Triangle3D

pub fn test_point(&self, p: Point3D) -> PointInTriangle

Tests whether a point—which is assumed to be within the plane of the triangle—is inside the triangle.

Explanation

This is not something I made up. I found it here and also in other places. (the explanation is here not as novelty)

If a triangle in 3 dimentions as vertices $\vec{A}$, $\vec{B}$, and $\vec{C}$, then any point in the plane of the triangle can be described by the following equation:

\vec{P} = \vec{A} + \alpha \left(\vec{B}-\vec{A}\right)+\beta \left(\vec{C}-\vec{A}\right)

Or, after stating that $\hat{e}_1=\left(\vec{B}-\vec{A}\right)$ and $\hat{e}_2 = \left(\vec{C}-\vec{A}\right)$ and rearranging, we can express the same equation as:

\vec{P} - \vec{A}=  \alpha \hat{e}_1+\beta \hat{e}_2

Since this is a 3D equation, it is actually three equations and there are only two incognita ($\alpha$ and $\beta$). So, we can multiply it by $\hat{e}_1$ and $\hat{e}_2$ and produce a system of equations that we can solve. After doing this, the equation above becomes:

\hat{e}_1\left(\vec{P} - \vec{A}\right)=  \alpha \hat{e}_1\hat{e}_1+\beta \hat{e}_1\hat{e}_2

and

\hat{e}_2\left(\vec{P} - \vec{A}\right)=  \alpha \hat{e}_1\hat{e}_2+\beta \hat{e}_2\hat{e}_2

Which can be represented in matrix form as follows:

\begin{Bmatrix}\hat{e}_1\left(\vec{P} - \vec{A}\right)\\\hat{e}_2\left(\vec{P} - \vec{A}\right)\end{Bmatrix}=\begin{pmatrix}\hat{e}_1\hat{e}_1&\hat{e}_1\hat{e}_2\\\hat{e}_1\hat{e}_2&\hat{e}_2\hat{e}_2\end{pmatrix}\begin{Bmatrix}\alpha\\\beta\end{Bmatrix}

What is left from here is to interpret the values of $\alpha$ and $\beta$ from the equations, which allows us to position the point within the triangle.

pub fn get_edge_index_from_segment(&self, s: &Segment3D) -> Option<usize>

Gets the edge number corresponding to a Segment3D in the Triangle.

pub fn get_edge_index_from_points(&self, a: Point3D, b: Point3D) -> Option<usize>

Retreives a certain edge based on two vertices

pub fn has_vertex(&self, p: Point3D) -> bool

Checks if a Triangle3D contains a certain Vertex in the same position as a Point3D p

pub fn compare(&self, t: &Triangle3D) -> bool

Checks whether three Triangle3D objects are made of he same vertices.

Vertices do not need to be in the same order, necessarily… So, check if the three vertices are in the other triangle.

pub fn basic_intersection(
&self,
ray: &Ray3D,
_o_error: Point3D,
_d_error: Point3D
) -> Option<(Point3D, f64, f64)>

Intersects a Ray3D in local coordinates with the Triangle3D. Returns the a Point3D of intersection and the values u and v, indicating the parametric coordinates of the point of intersection

pub fn id(&self) -> &'static str

The name of the figure. Useful for debugging.

pub fn bounds(&self) -> BBox3D

Gets a BBox3D bounding the object, in local coordinates

pub fn transform(&self) -> &Option<Rc<Transform>>

Borrows the Transform

pub fn intersect_local_ray(
&self,
ray: &Ray3D,
o_error: Point3D,
d_error: Point3D
) -> Option<IntersectionInfo>

Intersects an object with a [Ray3D] (IN LOCAL COORDINATES) traveling forward, returning the distance t and the normal Vector3D at that point. If the distance is negative (i.e., the object is behind the plane), it should return None.

The steps should be:

Calculate the point of intersection

Calculate normal and side of intersection

Calculate (u,v)

Calculate first derivative (i.e., dp_du, dp_dv)

Calculate second derivative (i.e., d2p_duu, d2p_dvv, d2p_duv)

Call IntersectionInfo::new(ray,p, u, v, dpdu, dpdv, d2p_duu, d2p_dvv, d2p_duv)

pub fn simple_intersect_local_ray(
&self,
ray: &Ray3D,
o_error: Point3D,
d_error: Point3D
) -> Option<Point3D>

Like intersect_local_ray but simplified because there is not need for calcuating the paramtrisized elements

pub fn intersect(&self, ray: &Ray3D) -> Option<IntersectionInfo>

Intersects an object with a [Ray3D] (IN WORLD COORDINATES) traveling forward, returning a detailed IntersectionInfo about the intersaction .

pub fn simple_intersect(&self, ray: &Ray3D) -> Option<Point3D>

Intersects an object with a [Ray3D] (IN WORLD COORDINATES) traveling forward, returning the point of intersection, if any.

pub fn world_bounds(&self) -> BBox3D

Gets a BBox3D bounding the object, in world’s coordinates.