Struct geometry3d::Triangulation3D

pub struct Triangulation3D {
    triangles: Vec<TriPiece>,
    n_valid_triangles: usize,
}

A set of Triangle3D that, together, cover completely a Polygon3D.

Fields
triangles: Vec<TriPiece>

The triangles

n_valid_triangles: usize

The number of valid triangles.

Check whether triangles are valid or not before using them. Invalidating triangles help accelerate the algorithms in this module.

Implementations

impl Triangulation3D

pub fn new() -> Triangulation3D

Creates a new empty Triangulation3D

pub fn with_capacity(i: usize) -> Triangulation3D

Creates a new empty Triangulation3D with a certain capacity

pub fn get_trilist(&self) -> Vec<Triangle3D>

pub fn mesh_polygon(
poly: &Polygon3D,
max_area: f64,
max_aspect_ratio: f64
) -> Result<Triangulation3D, String>

Triangulates a Polygon3D using the simple ear clipping algorithm and then progresively refines the mesh in order to get a relatively healthy set of Delaunay triangles

fn mark_neighbourhouds(&mut self) -> Result<(), String>

Loops over all Triangle3D, setting the neighbours. THis is quite slow, but does the job.

pub fn from_polygon(poly: &Polygon3D) -> Result<Triangulation3D, String>

Triangulates a Polygon3D without refining it, using the simple ear clipping algorithm.

pub fn n_triangles(&self) -> usize

Returns the number of Triangle3D in the Triangulation3D including invalid ones.

pub fn n_valid_triangles(&self) -> usize

Returns the number of valid Triangle3D in the Triangulation3D

fn invalidate(&mut self, i: usize) -> Result<(), String>

fn get_first_invalid(&self, start: usize) -> Option<usize>

fn mark_as_neighbours(
&mut self,
i1: usize,
edge_1: Edge,
i2: usize
) -> Result<(), String>

Marks two triangles as neighbours

pub fn push(
&mut self,
vertex_a: Point3D,
vertex_b: Point3D,
vertex_c: Point3D,
last_added: usize
) -> Result<usize, String>

fn get_opposite_vertex(
&self,
triangle: &Triangle3D,
segment: Segment3D
) -> Result<Point3D, String>

Gets the vertex that is opposite to a Segment3D in a Triangle3D

fn get_flipped_aspect_ratio(&self, index: usize, edge: Edge) -> Option<f64>

Returns the aspect ration that would result after flipping the diagonal of Triangle3D (located at index)

fn flip_diagonal(&mut self, index: usize, edge: Edge)

fn split_edge(
&mut self,
triangle_index: usize,
edge_to_split: Edge,
p: Point3D
) -> Result<(), String>

Adds a point to an edge of a Triangle3D, splitting it and its neighbour

fn split_triangle(&mut self, i: usize, point: Point3D) -> Result<(), String>

Adds a point to the interior of a Triangle3D, splitting it into three

fn restore_delaunay(&mut self, max_aspect_ratio: f64)

Runs through the Triangulation3D checking if it is worth flipping diagonals. I.e., if flipping the diagonal would make the aspect ratios become better

fn add_point_to_triangle(
&mut self,
index: usize,
point: Point3D,
p_location: PointInTriangle
) -> bool

Adds a point on a specific Triangle3D. Returns a bool, indicating whether any change was carried out

fn add_point(&mut self, point: Point3D) -> Result<bool, String>

Finds the location of a Point3D within the triangulation and inserts it. Returns a bool indicating if the addition of the new Point3D produced any change

fn refine(&mut self, max_area: f64, max_aspect_ratio: f64) -> Result<(), String>

Refines a Triangulation3D. This only works if all then Triangle3D have their vertices in the same order, meaning that their normal points in the same direction.