Struct geometry3d::Vector3D

pub struct Vector3D {
    pub x: f64,
    pub y: f64,
    pub z: f64,
}

A 3-dimensional Vector

Fields
x: f64

the X component

y: f64

the Y component

z: f64

the Z component

Implementations

impl Vector3D

pub fn new(x: f64, y: f64, z: f64) -> Vector3D

Creates a new vector.

pub fn abs(&self) -> Self

Returns a new vector, containing the absolute values of each component of self

pub fn as_point3d(&self) -> Point3D

Transforms a Vector3D into a Point3D

pub fn get_perpendicular(&self) -> Result<Self, String>

Gets a normalized Vector3D that is perpendicular to self

pub fn compare(&self, p: Vector3D) -> bool

Checks whether a vector is equal (or close) to another vector

pub fn cross(&self, v: Vector3D) -> Vector3D

Cross product between two vectors

pub fn length(&self) -> f64

Calculates the length of a vector

pub fn length_squared(&self) -> f64

Calculates the length of a vector squared.

Often using the squared of the length is the same as using the length of a vector (e.g., when comparing distances, the squared lengths of vectors are sorted in the same order as the lengths themselves). In such cases, using length_squared() is preferred because it is faster

pub fn normalize(&mut self)

Normalizes a Vector, mutating it.

pub fn get_normalized(&self) -> Vector3D

Returns a normalized version of a vector

pub fn is_zero(&self) -> bool

Checks whether a vector is Zero

pub fn is_parallel(&self, v: Vector3D) -> bool

Checks if two vectors are parallel

pub fn is_same_direction(&self, v: Vector3D) -> bool

Checks if two vectors are parallel and follow the same direction.