Crate geometry3d

This library has several 3D Geometry functionalities. E.g.,

Points

Vectors

Segments

Transforms

Triangles

Triangulation (e.g., sort of meshing… but not very high quality for finite element methods, probably)

Polygons

Ray/Primitive intersection algorythms (not fast enough for rendering, though, as it uses dynamic dispatch)

Re-exports

pub use transform::Transform;

Modules

bbox3d 🔒

cylinder3d 🔒

disk3d 🔒

distant_source3d 🔒

intersection

Intersection object

loop3d 🔒

plane3d 🔒

point3d 🔒

polygon3d 🔒

ray3d 🔒

round_error

Rounding Error accumulation module, based on PBRT

segment3d 🔒

sphere3d 🔒

transform

Transforms

triangle3d 🔒

triangulation3d 🔒

utils 🔒

vector3d 🔒

Macros

gamma

Calculates the gamma for error analysis

Structs

BBox3D

An Axis-aligned Bounding Box, represented by two corners (i.e., by two Point3D objects)

Cylinder3D

A Cylinder of radius radius aligned witht the Z axis, starting at a Z value of zmin and ending at zmax.

Disk3D

A disk of radius radius whose normal points in the normal direction, located at a height height. Can have a

DistantSource3D

Represents a solid angle pointing from any Point3D in the scene towards a certain direction (i.e., a Vector3D). So, this is not really a surface because it is at an infinite distance from any point.

Loop3D

A set of Point3D in sequence, forming a closed loop. It has some particularities.

Plane3D

Describes a plane based on a Point3D (P) contained in the plane and a Vector3D (N) normal to the plane.

Point3D

A very simple implementation of a 3D-Point

Polygon3D

A 3-dimensional Polygon which can contain holes.

Ray3D

A simple Ray structure, consiting of an origin (i.e., Point3D) and a direction (i.e., a Vector3D)

Segment3D

An imaginary line starting at one Point3D and ending on another Point3D.

Sphere3D

A Full or Partial sphere in three dimensions, centered at the origin

Triangle3D

A simple 3-dimentional Triangle.

Triangulation3D

A set of Triangle3D that, together, cover completely a Polygon3D.

Vector3D

A 3-dimensional Vector

Enums

BBoxAxis

Represents one dimension of the BBox3D

PointInTriangle

Represents the position of a Point3D in a Triangle3D.

Constants

PI 🔒

Type Definitions

Float 🔒

RefCount 🔒