Struct geometry3d::polygon3d::Polygon3D

pub struct Polygon3D {
    outer: Loop3D,
    inner: Vec<Loop3D>,
    area: f64,
    normal: Vector3D,
}

Fields
outer: Loop3Dinner: Vec<Loop3D>area: f64normal: Vector3D
Implementations

impl Polygon3D

pub fn new(outer: Loop3D) -> Result<Polygon3D, String>

Creates a new Loop3D without any holes

pub fn normal(&self) -> Vector3D

Retrieves the normal to the Polygon

pub fn outer_centroid(&self) -> Point3D

Calculates the average of the Point3D in the Outer Loop3D

pub fn outer(&self) -> &Loop3D

Borrows the Outer Loop3D

pub fn clone_outer(&self) -> Loop3D

Clones the Outer Loop3D

pub fn n_inner_loops(&self) -> usize

Counts the inner Loop3D

pub fn inner(&self, i: usize) -> Result<&Loop3D, String>

Borrows an inner Loop3D

pub fn test_point(&self, p: Point3D) -> Result<bool, String>

Checks whether a Point3D is inside the Polygon3D

pub fn cut_hole(&mut self, hole: Loop3D) -> Result<(), String>

Makes a hole shaped as a Loop3D in the Polygon3D.

pub fn area(&self) -> f64

Gets the area of the Polygon3D

pub fn get_closed_loop(&self) -> Loop3D

Gets a single Loop3D representing the same Polygon3D geometry, but without any holes

This is a pretty terible algorithm… but I doubt it is the bottleneck in my applications.

It is also not 100% guaranteed to work all the time, but it has proven to be pretty robust for relatively normal geometries.

pub fn contains_segment(&self, s: &Segment3D) -> bool

Checks whether the Polygon3D contains a certain Segment3D in any of its loops