Struct geometry3d::BBox3D

pub struct BBox3D {
    pub min: Point3D,
    pub max: Point3D,
}

An Axis-aligned Bounding Box, represented by two corners (i.e., by two Point3D objects)

Fields
min: Point3D

The lower corner

max: Point3D

Higher corner

Implementations

impl BBox3D

pub fn new(a: Point3D, b: Point3D) -> Self

Creates a new BBox3D from two points. The points will be all scrambled up in order to find the minimum x, y and z.

pub fn from_point(p: Point3D) -> Self

Creates a BBox3D that contains a single point, thus its min and its max values are equal

pub fn from_union_point(bbox: &Self, pt: Point3D) -> Self

Creates a new BBox3D that contains an existing BBox3D and a Point3D

pub fn from_union(bbox1: &Self, bbox2: &Self) -> Self

Creates a new BBox3D that contains two existing BBox3D

pub fn from_intersection(bbox1: &Self, bbox2: &Self) -> Self

Creates a new BBox3D that contains the intersection of two existing BBox3D

pub fn overlaps(&self, other: &Self) -> bool

Checks if a BBox3D overlaps with another one

pub fn point_inside(&self, pt: Point3D) -> bool

Checks if a Point3D is inside this BBox3D

pub fn point_inside_exclusive(&self, pt: Point3D) -> bool

Checks if a Point3D is inside this BBox3D but any Point3D located in the upper boundary is considered to be outside

pub fn max_extent(&self) -> BBoxAxis

Returns the dimension in which the BBox3D is longest

pub fn surface_area(&self) -> f64

Returns the total surface area of all the walls in this BBox3D

pub fn intersect(&self, ray: &Ray3D, inv_dir: &Vector3D) -> bool

Checks if a ray intersects a BBox3D

This function intersects the ray with 6 axis-aligned planes: 2 whose normal is Vector3D(1., 0., 0.), 2 whose normal is Vector3D(0., 1., 0.) and 2 whose normal is Vector3D(0., 0., 1.). That is to way, we calculate the time that the ray spends within the boundaries of the BBox3D in each dimension, and then find the intersection of those times. If there is no intersection, then the ray was never inside of the box.

For example, lets say that we realize that a ray spends t=[0., 3.] inside of the boundaries of the BBox3D in the X dimension; t=[0.2, 5.] in the Y dimension; and t=[-20., 1.] in the Z dimension. The intersection of these ranges is t=[0.2, 1.], which works and thus the ray went through the box.

Now let’s take an example in which the ray does not go through the BBox3D. For example, one in which a ray spends t=[0., 3.] inside of the boundaries of the BBox3D in the X dimension; t=[0.2, 5.] in the Y dimension; and t=[6., 12.] in the Z dimension. The intersection of these ranges is t=[6, 3.], which does not make any sense and thus the ray never went inside of the BBox3D.

Example

use geometry3d::{BBox3D, Vector3D, Point3D, Ray3D};
let min = Point3D::new(0., 0., 0.);
let max = Point3D::new(1., 1., 1.);
let bbox = BBox3D::new(min, max);
let ray = Ray3D {
    origin: Point3D::new(0.5, 0.5, 0.5),
    direction: Vector3D::new(0., 0., 1.)
};
let inv_dir = Vector3D::new(1./ray.direction.x, 1./ray.direction.y, 1./ray.direction.z);    
assert!(bbox.intersect(&ray, &inv_dir));